+++ bin/apis/linetv-list-station-request.js
@@ -16,7 +16,7 @@ class LINETvListStationRequest extends _linetvRequest.default {
   }
   /**
     * @param {number} channelId
-    * @param {string} channelCode
+    * @param {string} countryCode
     * @return {AxiosResponse<LINETvListStationResponseData>}
     */