+++ bin/typedef/linetv-live-response-data.js
@@ -1,0 +1,7 @@
+/**
+ * @typedef LINETvLiveResponseData
+ * @property {{ message: string; code: number }} header
+ * @property {{ lives: Array<LINETvLiveData> }} body
+ */
+"use strict";
+//# sourceMappingURL=linetv-live-response-data.js.map